backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# LangChain & LangGraph
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.60
langgraph-checkpoint-postgres==2.0.11

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis==5.2.1

# HTTP & API Clients
httpx==0.28.1
aiohttp==3.11.11

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Utilities
tenacity==9.0.0
email-validator==2.3.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
aiosqlite==0.22.1
